requirements.txt
pandas
numpy
requests
fuzzywuzzy
python-Levenshtein
openai
python-dotenv
markdown2
pdfkit
streamlit
